finished GUI

+++ src/drivers/mouse.cpp
@@ -29,7 +29,7 @@ void MouseEventHandler::OnMouseDown(uint8_t button){
 void MouseEventHandler::OnMouseUp(uint8_t button){
     
 }
-void MouseEventHandler::OnMouseMove(int8_t x, int8_t y){
+void MouseEventHandler::OnMouseMove(int x, int y){
     
 }
 
@@ -91,7 +91,7 @@ uint32_t MouseDriver::HandleInterrupt(uint32_t esp){
     if(offset ==0){
         if(buffer[1] !=0 || buffer[2] != 0 ){
             
-            handler->OnMouseMove(buffer[1], -buffer[2]);
+            handler->OnMouseMove(((int8_t)buffer[1]), - ((int8_t)buffer[2]));
             
         }
